% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_s3.R
\name{install_s3}
\alias{install_s3}
\title{Install packages from an s3 buckey}
\usage{
install_s3(bucket, package, awsKey, secretKey, deleteFiles = TRUE)
}
\arguments{
\item{bucket}{The name of the s3 where your packages are stored}

\item{package}{The name of your package (the folder name in the s3 bucket)}

\item{awsKey}{Your awsKey}

\item{secretKey}{Your aws secret key}

\item{deleteFiles}{Defaults to TRUE, all temp files created during installation will be deleted}
}
\description{
This function allows you to install R packages stored in an s3 bucket.
}
